Support building PAL region ROM
The PAL version of the ROM is undumped as of yet, so the ROM produced
here is speculative in nature. Nevertheless, it should work with PAL
CICs and given how constrained the changes are, there's a nonzero chance
it matches the true PAL ROM.

The C model is updated to be universal for convenience. The deviation in
behavior from real PIF ROMs in noted in comments.